Refactoring: Split project page into a block component for efficiency
